Use generated ids in client test
